+++ src/Readuct/App/Tasks/NtOptimization2Task.h
@@ -4,19 +4,15 @@
  *            Copyright ETH Zurich, Laboratory of Physical Chemistry, Reiher Group.\n
  *            See LICENSE.txt for details.
  */
-#ifndef READUCT_AFIROPTIMIZATIONTASK_H_
-#define READUCT_AFIROPTIMIZATIONTASK_H_
+#ifndef READUCT_NTOPTIMIZATION2TASK_H_
+#define READUCT_NTOPTIMIZATION2TASK_H_
 
 /* Readuct */
 #include "Tasks/Task.h"
 /* Scine */
-#include <Utils/GeometryOptimization/AfirOptimizer.h>
-#include <Utils/GeometryOptimization/GeometryOptimization.h>
+#include <Utils/GeometryOptimization/NtOptimizer2.h>
 #include <Utils/IO/ChemicalFileFormats/XyzStreamHandler.h>
-#include <Utils/Optimizer/GradientBased/Bfgs.h>
-#include <Utils/Optimizer/GradientBased/Lbfgs.h>
-#include <Utils/Optimizer/GradientBased/SteepestDescent.h>
-#include <Utils/Optimizer/HessianBased/NewtonRaphson.h>
+#include <Utils/IO/Yaml.h>
 /* External */
 #include <boost/exception/diagnostic_information.hpp>
 #include <boost/filesystem.hpp>
@@ -28,20 +24,20 @@
 namespace Scine {
 namespace Readuct {
 
-class AfirOptimizationTask : public Task {
+class NtOptimization2Task : public Task {
  public:
   /**
-   * @brief Construct a new AfirOptimizationTask.
+   * @brief Construct a new NtOptimization2Task.
    * @param input  The input system names for the task.
    * @param output The output system names for the task.
    * @param logger The logger to/through which all text output will be handled.
    */
-  AfirOptimizationTask(std::vector<std::string> input, std::vector<std::string> output, std::shared_ptr<Core::Log> logger = nullptr)
+  NtOptimization2Task(std::vector<std::string> input, std::vector<std::string> output, std::shared_ptr<Core::Log> logger = nullptr)
     : Task(std::move(input), std::move(output), std::move(logger)) {
   }
 
   std::string name() const override {
-    return "AFIR Optimization";
+    return "NT2 Optimization";
   }
 
   bool run(SystemsMap& systems, Utils::UniversalSettings::ValueCollection taskSettings, bool testRunOnly = false) const final {
@@ -57,37 +53,14 @@ class AfirOptimizationTask : public Task {
 
       // Check system size
       if (calc->getStructure()->size() == 1) {
-        throw std::runtime_error("Cannot calculate AFIR for monoatomic systems.");
+        throw std::runtime_error("Cannot calculate NT2 optimization for monoatomic systems.");
       }
     }
 
     // Generate optimizer
-    auto optimizertype = taskSettings.extract("optimizer", std::string{"BFGS"});
-    std::transform(optimizertype.begin(), optimizertype.end(), optimizertype.begin(), ::toupper);
-    std::shared_ptr<Utils::AfirOptimizerBase> optimizer;
-    if (optimizertype == "LBFGS") {
-      auto tmp = std::make_shared<Utils::AfirOptimizer<Utils::Lbfgs>>(*calc);
-      tmp->optimizer.useTrustRadius = true;
-      tmp->optimizer.trustRadius = 0.1;
-      optimizer = std::move(tmp);
-    }
-    else if (optimizertype == "BFGS") {
-      auto tmp = std::make_shared<Utils::AfirOptimizer<Utils::Bfgs>>(*calc);
-      tmp->optimizer.useTrustRadius = true;
-      tmp->optimizer.trustRadius = 0.1;
-      optimizer = std::move(tmp);
-    }
-    else if (optimizertype == "SD" || optimizertype == "STEEPESTDESCENT") {
-      auto tmp = std::make_shared<Utils::AfirOptimizer<Utils::SteepestDescent>>(*calc);
-      optimizer = std::move(tmp);
-    }
-    else {
-      throw std::runtime_error("Unknown Optimizer requested for AFIR, available are: SD, BFGS and LBFGS!");
-    }
-
+    auto optimizer = std::make_unique<Utils::NtOptimizer2>(*calc);
     // Read and delete special settings
     bool stopOnError = stopOnErrorExtraction(taskSettings);
-
     // Apply user settings
     auto settings = optimizer->getSettings();
     settings.merge(taskSettings);
@@ -95,10 +68,6 @@ class AfirOptimizationTask : public Task {
       settings.throwIncorrectSettings();
     }
     optimizer->setSettings(settings);
-    if (!testRunOnly && !Utils::GeometryOptimization::settingsMakeSense(*optimizer)) {
-      throw std::logic_error("The given calculator settings are too inaccurate for the given convergence criteria of "
-                             "this optimization Task");
-    }
 
     // If no errors encountered until here, the basic settings should be alright
     if (testRunOnly) {
@@ -108,30 +77,30 @@ class AfirOptimizationTask : public Task {
     // Add observer
     // Trajectory stream
     using Writer = Utils::XyzStreamHandler;
-    const std::string& outputSystem = (!_output.empty() ? _output.front() : _input.front());
+    const std::string& outputSystem = ((!_output.empty()) ? _output[0] : _input[0]);
     boost::filesystem::path dir(outputSystem);
     boost::filesystem::create_directory(dir);
-    boost::filesystem::path trjfile(outputSystem + ".afir.trj.xyz");
+    boost::filesystem::path trjfile(outputSystem + ".nt.trj.xyz");
     std::ofstream trajectory((dir / trjfile).string(), std::ofstream::out);
     double oldEnergy = 0.0;
     Eigen::VectorXd oldParams;
-    auto cout = _logger->output;
     auto func = [&](const int& cycle, const double& energy, const Eigen::VectorXd& params) {
       if (oldParams.size() != params.size()) {
         oldParams = params;
       }
       if (cycle == 1) {
-        cout.printf("%7s %16s %16s %16s %16s\n", "Cycle", "Energy", "Energy Diff.", "Step RMS", "Max. Step");
+        printf("%7s %16s %16s %16s %16s\n", "Cycle", "Energy", "Energy Diff.", "Step RMS", "Max. Step");
       }
       auto diff = (params - oldParams).eval();
-      cout.printf("%7d %+16.9f %+16.9f %+16.9f %+16.9f\n", cycle, energy, energy - oldEnergy,
-                  sqrt(diff.squaredNorm() / diff.size()), diff.cwiseAbs().maxCoeff());
+      printf("%7d %+16.9f %+16.9f %+16.9f %+16.9f\n", cycle, energy, energy - oldEnergy,
+             sqrt(diff.squaredNorm() / diff.size()), diff.cwiseAbs().maxCoeff());
       oldEnergy = energy;
       oldParams = params;
       auto structure = calc->getStructure();
       Writer::write(trajectory, *structure, std::to_string(energy));
     };
     optimizer->addObserver(func);
+    auto cout = _logger->output;
 
     // Run optimization
     auto structure = calc->getStructure();
@@ -140,44 +109,39 @@ class AfirOptimizationTask : public Task {
       cycles = optimizer->optimize(*structure, *_logger);
     }
     catch (...) {
-      Writer::write(trajectory, *calc->getStructure());
       trajectory.close();
-      _logger->error << "AFIR Optimization failed with error!" << Core::Log::endl;
       if (stopOnError) {
         throw;
       }
-      _logger->error << boost::current_exception_diagnostic_information() << Core::Log::endl;
+      // check if thrown exception corresponds to no actual calculation failures but simply no guess found
+      // --> no thrown exception intended in this case if stopOnError equals false
+      std::string noGuess = "No transition state guess was found in Newton Trajectory scan";
+      size_t found = boost::current_exception_diagnostic_information().find(noGuess);
+      if (found == std::string::npos) { // did not find harmless exception -> throw
+        throw;
+      }
+      cout << Core::Log::endl << "    No TS guess found in NT2 scan." << Core::Log::endl << Core::Log::endl;
       return false;
     }
+
     trajectory.close();
 
-    int maxiter = settings.getInt("convergence_max_iterations");
-    if (maxiter > cycles) {
-      cout << Core::Log::endl
-           << "    Converged after " << cycles << " iterations." << Core::Log::endl
-           << Core::Log::endl;
-    }
-    else {
-      cout << Core::Log::endl
-           << "    Stopped after " << maxiter << " iterations." << Core::Log::endl
-           << Core::Log::endl;
-      if (stopOnError) {
-        throw std::runtime_error("Problem: AFIR optimization did not converge.");
-      }
-    }
+    cout << Core::Log::endl
+         << "    Found TS guess after " << cycles << " iterations." << Core::Log::endl
+         << Core::Log::endl;
 
     // Print/Store results
+    systems[outputSystem] = calc;
     boost::filesystem::path xyzfile(outputSystem + ".xyz");
     std::ofstream xyz((dir / xyzfile).string(), std::ofstream::out);
     Writer::write(xyz, *(calc->getStructure()));
     xyz.close();
-    systems[outputSystem] = std::move(calc);
 
-    return cycles < maxiter;
+    return true;
   }
 };
 
 } // namespace Readuct
 } // namespace Scine
 
-#endif // READUCT_AFIROPTIMIZATIONTASK_H_
+#endif // READUCT_NTOPTIMIZATION2TASK_H_